Sync pipeline orchestration: wire up pipeline start/stop operations
A unique pipeline_run_id is created if the pipeline IR is configured for runtime substitution. If pipeline_run_id is already populated in the IR, it is used directly -- this is for the case where client already knows which pipeline run to operate upon OR wants to substitute pipeline_run_id on the client side.

Other changes:
* Updated pipeline_ops tests to handle both sync and async pipelines.
* Use auto_attribs=True for attrs classes (mostly cosmetic change).
* Change how pipeline-level context name is created.

Upcoming:
* Updating sync pipeline state at the end of a run.
* Handling service jobs in sync pipelines.
PiperOrigin-RevId: 359368878